Rename workflow package to workflowservice
